pyproject.toml
[project]
name = "carbon-forecasting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "diagrams>=0.25.1",
    "fastparquet>=2025.12.0",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
